test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
hacking<0.11,>=0.10.0
coverage>=3.6 # Apache-2.0
discover # BSD
doc8 # Apache-2.
fixtures>=1.3.1 # Apache-2.0/BSD
requests-mock>=0.7.0 # Apache-2.0
mock>=1.2 # BSD
Babel>=1.3 # BSD
sphinx!=1.2.0,!=1.3b1,<1.3,>=1.1.2 # BSD
testrepository>=0.0.18 # Apache-2.0/BSD
testtools>=1.4.0 # MIT
